requirements.txt
django~=4.1.2
djangorestframework~=3.14.0
psycopg2
django-model-utils~=4.2.0
gunicorn
dj-config-url~=0.1.1
django-filter
drf-extensions